Pass alpha structure by reference, with appropriate parameters
The alpha structure was being passed to vc_dispmanx_element_add
by value, even though the function expected a pointer. No
visible output was produced in Raspbian Jessie, and setting a
non-zero opacity caused a crash. This sets up the structure for
a semi-transparent full screen display and then passes a pointer.